Change text box coordinates and dimensions to integer-only input
Updated TextBox model to use int instead of float for x, y, width, and height fields. Modified validators in app_config to enforce integer validation. Updated text_boxes_tab to use safe_int() instead of safe_float(). Updated Excel repository to read integer values. Backward compatibility maintained for loading old float values which are converted to integers.

+++ config/app_config.py
@@ -279,10 +279,10 @@ class AppConfig:
                 columns=[
                     TableColumnConfig("Select", widget_type="checkbox"),
                     TableColumnConfig("ID", validator=lambda x: int(x) > 0 if x else False),
-                    TableColumnConfig("X", validator=lambda x: float(x) >= 0 if x else False),
-                    TableColumnConfig("Y", validator=lambda x: float(x) >= 0 if x else False),
-                    TableColumnConfig("Width", validator=lambda x: float(x) > 0 if x else False),
-                    TableColumnConfig("Height", validator=lambda x: float(x) > 0 if x else False)
+                    TableColumnConfig("X", validator=lambda x: int(x) >= 0 if x else False),
+                    TableColumnConfig("Y", validator=lambda x: int(x) >= 0 if x else False),
+                    TableColumnConfig("Width", validator=lambda x: int(x) > 0 if x else False),
+                    TableColumnConfig("Height", validator=lambda x: int(x) > 0 if x else False)
                 ],
                 min_rows=0
             )
